-- separated components and added messages mapping
