+++ src/frontend/src/redux/Message/messageActions.js
@@ -35,7 +35,6 @@ export const addUserAction = ({chatId, userId}) => (dispatch) => {
 }
 
 export const getUserChatsAction = (userId) => (dispatch) => {
-  console.log(userId)
   return http
     .get(`/api/chats/user/${userId}`)
     .then(res => {
