+++ src/main/java/com/danit/fs12/entity/chat/ChatRq.java
@@ -10,6 +10,6 @@ public class ChatRq {
   @NotNull(message = "User id should be specified.")
   private Long userId;
 
-  @NotNull(message = "Chat id should be specified.")
+  @NotNull(message = "ChatsList id should be specified.")
   private Long chatId;
 }
