+++ src/main/java/com/danit/fs12/entity/message/MessageRs.java
@@ -9,5 +9,6 @@ public class MessageRs {
   private Long userId;
   private Long chatId;
   private String text;
+  private String created_date;
 
 }
