+++ src/main/resources/data.sql
@@ -186,22 +186,22 @@ VALUES (date '2015-01-20', date '2015-01-20'),
 INSERT INTO USERS_CHATS (user_id, chat_id)
 VALUES (1, 1),
        (2, 1),
+       (1, 2),
        (3, 2),
-       (4, 2),
-       (5, 3),
-       (6, 3),
-       (7, 4),
-       (8, 4);
+       (1, 3),
+       (4, 3),
+       (1, 4),
+       (5, 4);
 
 INSERT INTO MESSAGES (created_date, modified_date, text, chat_id, user_id)
-VALUES (date '2015-01-20', date '2015-01-20', 'Message text', 1, 2),
+VALUES (date '2015-01-20', date '2015-01-20', 'Message text', 1, 1),
        (date '2015-01-20', date '2015-01-20', 'Message text2', 1, 2),
-       (date '2015-01-20', date '2015-01-20', 'Message text3', 2, 3),
+       (date '2015-01-20', date '2015-01-20', 'Message text3', 2, 1),
        (date '2015-01-20', date '2015-01-20', 'Message text4', 2, 3),
-       (date '2015-01-20', date '2015-01-20', 'Message text5', 2, 4),
+       (date '2015-01-20', date '2015-01-20', 'Message text5', 2, 1),
        (date '2015-01-20', date '2015-01-20', 'Message text6', 3, 5),
-       (date '2015-01-20', date '2015-01-20', 'Message text7', 3, 6),
-       (date '2016-10-01', date '2016-10-01', 'Message text8', 4, 7),
-       (date '2017-02-11', date '2017-02-11', 'Message text9', 4, 8),
-       (date '2017-02-21', date '2017-02-21', 'Message text10', 4, 7),
-       (date '2017-05-09', date '2017-05-09', 'Message text11', 4, 7);
+       (date '2015-01-20', date '2015-01-20', 'Message text7', 3, 1),
+       (date '2016-10-01', date '2016-10-01', 'Message text8', 4, 4),
+       (date '2017-02-11', date '2017-02-11', 'Message text9', 4, 1),
+       (date '2017-02-21', date '2017-02-21', 'Message text10', 4, 4),
+       (date '2017-05-09', date '2017-05-09', 'Message text11', 4, 1);
